chore: restructure query context
